+++ src/main/java/com/github/lombrozo/testnames/rules/NotUsesSpecialCharacters.java
@@ -27,7 +27,7 @@ package com.github.lombrozo.testnames.rules;
 import com.github.lombrozo.testnames.Complaint;
 import com.github.lombrozo.testnames.Rule;
 import com.github.lombrozo.testnames.Case;
-import com.github.lombrozo.testnames.WrongTestNameComplaint;
+import com.github.lombrozo.testnames.complaints.WrongTestNameComplaint;
 import java.util.Collection;
 
 /**
