+++ src/test/java/com/github/lombrozo/testnames/rules/AllTestsInPresentSimpleTest.java
@@ -24,12 +24,8 @@
 
 package com.github.lombrozo.testnames.rules;
 
-import com.github.lombrozo.testnames.Complaint;
-import com.github.lombrozo.testnames.WrongTestName;
-import java.util.Collection;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 /**
