+++ src/test/java/com/github/lombrozo/testnames/rules/CompositeTestPathRuleTest.java
@@ -24,7 +24,6 @@
 
 package com.github.lombrozo.testnames.rules;
 
-import com.github.lombrozo.testnames.WrongTestName;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
